refactor(ava/data): improve dataset
