+++ packages/ava/src/data/analysis/index.ts
@@ -1,2 +1,1 @@
 export { analyzeField } from './field';
-export type { FieldType, FieldInfo, StringFieldInfo, NumberFieldInfo, DateFieldInfo } from './field/types';
