+++ packages/ava/src/data/analysis/types.ts
@@ -1,0 +1,1 @@
+export type { FieldType, FieldInfo, StringFieldInfo, NumberFieldInfo, DateFieldInfo } from './field/types';
