+++ packages/ava/src/data/dataset/field/index.ts
@@ -1,0 +1,4 @@
+import Series from './series';
+import DataFrame from './dataFrame';
+
+export { Series, DataFrame };
