+++ packages/ava/src/data/dataset/index.ts
@@ -1,6 +1,1 @@
-import Series from './series';
-import DataFrame from './data-frame';
-import GraphData from './graph-data';
-
-export { Series, DataFrame, GraphData };
-export * from './types';
+export * from './field';
