+++ packages/ava/src/data/index.ts
@@ -1,2 +1,3 @@
 export * from './analysis';
+export * from './dataset';
 export * as DataTypes from './types';
